requirements.txt
Django==1.3.1
django-tastypie==0.9.10
mimeparse==0.1.3
python-dateutil==1.5
python-digest==1.7
wsgiref==0.1.2
